python/requirements.txt
pandas
numpy
phonenumbers
twilio
Pillow
rembgFlask